+++ packages/orm/angel_orm_test/lib/src/has_one_test.dart
@@ -115,7 +115,7 @@ void hasOneTests(FutureOr<QueryExecutor> Function() createExecutor,
   test('sets null on false subquery', () async {
     var legQuery = LegQuery()
       ..where!.id.equals(originalLeg!.idAsInt)
-      ..foot!.where!.legId.equals(originalLeg!.idAsInt + 1024);
+      ..foot.where!.legId.equals(originalLeg!.idAsInt + 1024);
     var legOpt = await (legQuery.getOne(executor));
     expect(legOpt.isPresent, true);
     legOpt.ifPresent((leg) {