+++ packages/orm/angel_orm_test/lib/src/models/leg.dart
@@ -1,7 +1,6 @@
 library angel3_orm_generator.test.models.leg;
 
 import 'package:angel3_migration/angel3_migration.dart';
-import 'package:angel3_model/angel3_model.dart';
 import 'package:angel3_orm/angel3_orm.dart';
 import 'package:angel3_serialize/angel3_serialize.dart';
 import 'package:optional/optional.dart';