+++ packages/orm/angel_orm_test/lib/src/models/world.dart
@@ -1,5 +1,4 @@
 import 'package:angel3_migration/angel3_migration.dart';
-//import 'package:angel3_model/angel3_model.dart';
 import 'package:angel3_serialize/angel3_serialize.dart';
 import 'package:angel3_orm/angel3_orm.dart';
 import 'package:optional/optional.dart';
@@ -9,6 +8,7 @@ part 'world.g.dart';
 @serializable
 @Orm(tableName: 'world')
 abstract class _World {
+  @primaryKey
   int? id;
 
   @Column()